requirements.txt
######## Unified project requirements ########

# Web
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1

# Database
psycopg2-binary>=2.9.9  # PostgreSQL драйвер (обновлено для Python 3.13)
SQLAlchemy>=2.0.23  # Обновлено для Python 3.13 (2.0.23 имеет проблемы, используйте 2.0.39+)

# HTTP/clients
requests==2.31.0

# T-Bank VoiceKit ставим отдельно в setup_env.bat (без deps),
# чтобы обойти конфликт его зависимостей с актуальными boto3/botocore.

# Audio processing
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
pandas>=2.2.0
openpyxl>=3.1.5
jsonschema>=4.21.0
# gRPC и protobuf (для T-Bank VoiceKit)
# Для Python 3.13 используйте последние версии (1.75+), которые имеют предкомпилированные wheels
grpcio>=1.75.0
grpcio-tools>=1.75.0
googleapis-common-protos>=1.63.0
# Протоколы для совместимости с Tinkoff VoiceKit
# ВНИМАНИЕ: protobuf 6.x может конфликтовать с некоторыми пакетами
# Если есть конфликты, используйте: protobuf>=5.26.1,<6.0.0
protobuf>=5.26.1,<7.0.0
google>=3.0.0
opuslib>=3.0.1

# Config and file watching
PyYAML>=6.0.0
watchdog>=3.0.0

# LLM client used by analyzer
openai>=1.0.0

# Utilities
tenacity>=9.0.0
PyJWT>=2.8.0  # Для авторизации в T-Bank gRPC API
python-dotenv>=1.0.0  # Для загрузки переменных окружения из .env

# Optional: scheduling if used in docs/README
APScheduler>=3.10.4


